debugging readReceipt

+++ jocr/src/main/java/ReceiptParser.java
@@ -2,6 +2,6 @@ package jocr.src.main.java;
 import org.bytedeco.javacpp.opencv_core.IplImage;
 
 public interface ReceiptParser {
-    abstract IplImage cropReceipt(String path, float[][] points);
+    abstract IplImage cropReceipt(String path, int[][] points);
     abstract String readReceipt(IplImage croppedImage);
 }
